Uploader/pyproject.toml
[project]
name = "uploader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "pillow>=11.0.0",
    "paddlepaddle-gpu==3.2.2",
    "paddlex[multimodal]>=3.0.0",
    "pymupdf>=1.27.2.3",
    "pypdfium2<5",
    "torch>=2.4.0",
    "transformers>=4.46.0",
    "accelerate>=0.34.0",
]

[tool.uv]
override-dependencies = [
    "pypdfium2<5",
    "nvidia-cudnn-cu11==9.1.0.70 ; sys_platform == 'linux'",
    "nvidia-nccl-cu11==2.20.5 ; sys_platform == 'linux'",
]

[[tool.uv.index]]
name = "paddlepaddle"
url = "https://www.paddlepaddle.org.cn/packages/stable/cu118/"
explicit = true

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
paddlepaddle-gpu = { index = "paddlepaddle" }
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
